Implement a protocol for reading data from the switch.
The commands for reading have PHY=102.
With REG=1, it reads values via MDIO. PHY and REG to read are set in the 2 data bytes.
With REG=2, it reads from EEPROM. The index to read is the first byte of data.
With REG=3, it reads from Commands_Total (RAM). The index to read is the first byte of data.
Returned is first the success byte, and then 4 bytes in the same format as commands [PHY, REG, DataByteHigh, DataByteLow].

Command with PHY=103 returns the success byte and then one byte with firmware revision number.

Command with PHY=104 returns the success byte and then 's' denoting this is a SwitchBlox Nano.